+++ src/js/Manager/ExportManager.js
@@ -48,7 +48,7 @@ export class ExportManager {
      */
     static async exportJson(model, options) {
 
-        let json = {version: 1, encrypted: false};
+        let json = {version: 2, encrypted: false};
         if(model.indexOf('passwords') !== -1) {
             json.passwords = await ExportManager._getPasswordsForExport(options.includeShared);
         }