+++ src/vue/Dialog/CreatePassword/CustomFields.vue
@@ -1,13 +1,12 @@
 <template>
     <div class="custom-fields" id="custom-fields">
-        <custom-field-form :field="field"
-                           :taken-names="getTakenNames"
-                           @deleted="deleteField"
-                           @updated="updateField"
-                           v-for="(field, index) in getFields"
-                           :key="index"
-                           :is-blank="field === undefined">
-            <hr v-if="index !== getFields.length-1">
+        <custom-field-form
+                :field="field"
+                @deleted="deleteField"
+                @updated="updateField"
+                v-for="(field, index) in getFields"
+                :key="index">
+            <hr v-if="index !== fields.length">
         </custom-field-form>
     </div>
 </template>
@@ -21,92 +20,44 @@
         components: {CustomFieldForm},
         props     : {
             fields: {
-                type: Object
+                type: Array
             }
         },
         data() {
             return {
-                showHiddenFields: SettingsManager.get('client.ui.custom.fields.show.hidden'),
-                customFields    : this.fields,
-                fieldPositions  : {}
+                showHidden  : SettingsManager.get('client.ui.custom.fields.show.hidden')
             };
         },
         computed  : {
             getFields() {
-                let fields     = [],
-                    fieldCount = 0;
+                let fields = [];
 
-                for(let name in this.customFields) {
-                    if(!this.customFields.hasOwnProperty(name)) continue;
-                    fieldCount++;
-                    if(!this.showHiddenFields && name.substr(0, 1) === '_') continue;
+                for(let i = 0; i < this.fields.length; i++) {
+                    let field = this.fields[i];
+                    if(field.type === 'data' && !this.showHidden) continue;
 
-                    let position = fields.length;
-                    if(this.fieldPositions.hasOwnProperty(name)) {
-                        position = this.fieldPositions[name];
-                    } else {
-                        this.fieldPositions[name] = position;
-                    }
-                    fields.push(
-                        {
-                            name,
-                            position,
-                            type : this.customFields[name].type,
-                            value: this.customFields[name].value
-                        }
-                    );
+                    field.id = i;
+                    field.blank = false;
+                    fields.push(field);
                 }
 
-                if(fieldCount < 20) fields.push(undefined);
-                return Utility.sortApiObjectArray(fields, 'position');
-            },
-            getTakenNames() {
-                return Object.keys(this.customFields);
+                if(fields.length < 20) {
+                    fields.push({label: '', type: 'text', value: '', id: fields.length, blank: true});
+                }
+
+                return fields;
             }
         },
         methods   : {
             updateField($event) {
-                let fields       = Utility.cloneObject(this.customFields),
-                    originalName = $event.originalName,
-                    name         = $event.name;
-
-                if(originalName !== name) {
-                    if(this.fieldPositions.hasOwnProperty(originalName)) {
-                        this.fieldPositions[name] = this.fieldPositions[originalName];
-                        delete this.fieldPositions[originalName];
-                    }
-
-                    if(originalName !== '') delete fields[originalName];
-                }
-
-                fields[name] = {
-                    type : $event.type,
-                    value: $event.value
-                };
-                this.customFields = fields;
+                let fields = Utility.cloneObject(this.fields);
+                fields[$event.id] = {label: $event.label, type: $event.type, value: $event.value};
+                this.$emit('updated', fields);
             },
             deleteField($event) {
-                let fields = Utility.cloneObject(this.customFields);
-                delete fields[$event];
-                this.customFields = fields;
-            }
-        },
-        watch     : {
-            fields() {
-                if(JSON.stringify(this.fields) !== JSON.stringify(this.customFields)) {
-                    this.customFields = this.fields;
-                }
-            },
-            customFields() {
-                let fields = {};
-                for(let name in this.customFields) {
-                    if(!this.customFields.hasOwnProperty(name) || !name.length) continue;
-                    let field = this.customFields[name];
-
-                    fields[name] = {type: field.type, value: field.value};
-                }
-
-                this.$emit('updated', fields);
+                let customFields = Utility.cloneObject(this.fields);
+                customFields.splice($event.id, 1);
+                this.$emit('updated', customFields);
             }
         }
     };