+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-#### 2019.3.0
+#### 2019.4.0 - Feature release
+- Added Nextcloud 16 support
+- Improved custom fields
+- Fixed double sse key generation
+- Fixed tags being not removed from passwords
+
+#### 2019.3.0 - Feature release
 - Improved legacy api password update
 - Improved import performance
 - Fixed deleting shared passwords issue